pyproject.toml
[project]
name = "aum"
version = "0.2.0"
description = "eDiscovery search engine with pluggable backends"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "prometheus-client>=0.20",
    "tika>=3.0",
    "elasticsearch>=8.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pyjwt>=2.9",
    "argon2-cffi>=23.1",
    "authlib>=1.3",
    "httpx>=0.27",
    "rich>=14.1.0",
]

[project.optional-dependencies]
embeddings = ["sentence-transformers>=3.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "mypy>=1.10",
    "httpx>=0.27",
]

[project.scripts]
aum = "aum.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aum"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
